(feat) Configured initial server setting
