+++ server/src/server.js
@@ -23,7 +23,7 @@ app.set('port', process.env.PORT || 4444);
 
 // now server running...
 app.listen(app.get('port'), function() {
-    console.log("Express started on http://localhost:" +
-        app.get('port') + '; press Ctrl-C to terminate.');
+  console.log('Express started on http://localhost:'+app.get('port'),
+              'press Ctrl-C to terminate.');
 });
